# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shop Mall API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: internal
  description: |
    The single source of truth for the buyer and admin APIs.
    All timestamps are RFC3339 UTC. Points, prices, stock, and quantities are integers.
    Public int64 values are serialized as decimal strings to preserve precision in JavaScript.
    Administrator authentication uses admin_access_token with a 30-minute lifetime and
    attributes HttpOnly; Secure; SameSite=Strict; Path=/api/admin/v1. csrf_token is a
    Secure; SameSite=Strict; Path=/api/admin/v1 cookie readable by the frontend, but it
    is never an identity credential. The server rotates csrf_token at login, requires
    every administrator write's X-CSRF-Token to exactly match the csrf_token cookie,
    clears both cookies with Max-Age=0 and the same attributes on logout, and accepts
    writes only from an allowed HTTPS Origin. Missing or invalid Origin and missing or
    mismatched CSRF values return 403 with a CSRF/Origin error. Login is the only
    administrator write that does not require an existing CSRF token, but it still
    validates Origin and rate limits by username and IP.
servers:
  - url: https://api.shop-mall.invalid
    description: Deployment replaces this logical server URL with PUBLIC_BASE_URL.
...
paths:
...
  /api/v1/products:
    get:
      tags: [products]
      operationId: listProducts
      summary: List products on sale
      description: Only products with status on_sale are returned. The category parameter restricts the result to one catalog key; keyword matches product names case-insensitively.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PublicProductPageSize'
        - $ref: '#/components/parameters/ProductCategory'
        - $ref: '#/components/parameters/ProductKeyword'
      responses:
        '200':
          description: Products currently on sale, ordered by id descending.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/categories:
    get:
      tags: [products]
      operationId: listCategories
      summary: List the product category catalog
      description: The static category catalog paired with the live on-sale product count of each category.
      security: []
      responses:
        '200':
          description: Category catalog in display order.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryInfo'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/products/{productId}:
    get:
      tags: [products]
      operationId: getProduct
      summary: Get a public product
      description: A product with status off_sale is intentionally indistinguishable from a missing product and returns 404.
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product detail.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/admin/v1/products:
    get:
      tags: [admin-products]
      operationId: adminListProducts
      summary: List products for administrators
      security:
        - AdminCookie: []
      x-required-permission: product:read
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/ProductStatus'
      responses:
        '200':
          $ref: '#/components/responses/ProductListSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [admin-products]
      operationId: adminCreateProduct
      summary: Create a product
      description: This administrator write requires an allowed HTTPS Origin and an X-CSRF-Token value that exactly matches the csrf_token cookie; missing or invalid Origin or CSRF values return 403.
      security:
        - AdminCookie: []
          CsrfCookie: []
          CsrfToken: []
      x-required-permission: product:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
      responses:
        '201':
          $ref: '#/components/responses/ProductSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/BusinessError'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/v1/products/{productId}:
    get:
      tags: [admin-products]
      operationId: adminGetProduct
      summary: Get a product for administrators
      security:
        - AdminCookie: []
      x-required-permission: product:read
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          $ref: '#/components/responses/ProductSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [admin-products]
      operationId: adminUpdateProduct
      summary: Update a product
      description: This administrator write requires an allowed HTTPS Origin and an X-CSRF-Token value that exactly matches the csrf_token cookie; missing or invalid Origin or CSRF values return 403.
      security:
        - AdminCookie: []
          CsrfCookie: []
          CsrfToken: []
      x-required-permission: product:write
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
      responses:
        '200':
          $ref: '#/components/responses/ProductSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessError'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Buyer JWT. The token is sent in the Authorization header and is never stored in an administrator cookie.
    AdminCookie:
      type: apiKey
      in: cookie
      name: admin_access_token
      description: Administrator JWT cookie with a 30-minute lifetime and HttpOnly; Secure; SameSite=Strict; Path=/api/admin/v1. It is cleared on logout or token invalidation and cannot be read by frontend JavaScript.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: Secure; SameSite=Strict; Path=/api/admin/v1 CSRF cookie. Frontend JavaScript may read it to copy the value into X-CSRF-Token; it is not an identity credential and is rotated at login.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on every administrator write except login. The value must exactly match the csrf_token cookie; missing or mismatched values return 403 with a CSRF error.
  headers:
    TraceId:
      description: Request trace identifier returned in the response body when the response uses the business envelope.
      required: true
      schema:
        type: string
        minLength: 1
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PublicProductPageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Public product, category, and search pages use 10 items by default.
...
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[1-9][0-9]*$'
        x-public-int64: true
        example: '1'
...
    ProductStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ProductStatus'
    ProductCategory:
      name: category
      in: query
      required: false
      description: Restrict the listing to one category catalog key; empty or absent means every category.
      schema:
        $ref: '#/components/schemas/CategoryKey'
    ProductKeyword:
      name: keyword
      in: query
      required: false
      description: Case-insensitive substring match against the product name.
      schema:
        type: string
        maxLength: 64
  responses:
...
    ProductSuccess:
      description: Product response.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductResponse'
    ProductListSuccess:
      description: Product list response.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductListResponse'
...
    Unauthorized:
      description: Authentication failed or token expired.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted. This 403 also covers missing or invalid Origin and missing or mismatched X-CSRF-Token/csrf_token values.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource does not exist or is not owned by the buyer.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BusinessError:
      description: Request is syntactically valid but violates business rules.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Request syntax or parameter format is invalid.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error. Internal details are not exposed.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, data, message, trace_id]
      properties:
        code:
          type: integer
          minimum: 1
          enum: [1001, 1002, 1003, 1004, 2001, 2002, 3001, 3002, 4001, 4002, 5001]
        data:
          type: 'null'
        message:
          type: string
        trace_id:
          type: string
          minLength: 1
...
    PageMeta:
      type: object
      required: [list, total, page, page_size]
      properties:
        list:
          type: array
          items: {}
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
...
    CategoryKey:
      type: string
      description: Stable category key from the catalog; empty means uncategorized.
      maxLength: 32
      example: digital
    CategoryInfo:
      type: object
      required: [id, name, product_count]
      properties:
        id:
          $ref: '#/components/schemas/CategoryKey'
        name:
          type: string
          description: Chinese display label of the category.
          example: 数码
        product_count:
          type: integer
          description: Live count of on-sale products in this category.
          minimum: 0
    Product:
      type: object
      required: [id, name, description, category, main_image, images, price_points, stock, status]
      properties:
        id:
          type: string
          pattern: '^[1-9][0-9]*$'
          x-public-int64: true
          example: '1'
        name:
          type: string
          maxLength: 128
        description:
          type: string
        category:
          $ref: '#/components/schemas/CategoryKey'
        main_image:
          type: string
          description: First gallery entry's public URL; kept for older clients. Empty when the product has no images.
        images:
          type: array
          description: Ordered gallery of public image URLs. Position 0 is the main image shown in listings and carts; blanks and duplicates never appear; at most 9 entries.
          maxItems: 9
          items:
            type: string
        price_points:
          type: string
          pattern: '^[1-9][0-9]*$'
          x-public-int64: true
          example: '100'
        stock:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductStatus:
      type: string
      enum: [on_sale, off_sale]
    ProductWriteRequest:
      type: object
      required: [name, price_points, stock]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
        category:
          description: Stable category key; empty or absent means uncategorized. Unknown keys are rejected with 422.
          $ref: '#/components/schemas/CategoryKey'
        images:
          type: array
          description: Full replacement gallery of previously uploaded image object keys, ordered (position 0 = main image). Blanks are dropped and duplicates collapse server-side; more than 9 entries is rejected with 422. An absent field on PATCH keeps the current gallery; an empty array clears it.
          maxItems: 9
          items:
            type: string
        price_points:
          type: string
          pattern: '^[1-9][0-9]*$'
          x-public-int64: true
          example: '100'
        stock:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/ProductStatus'
...
    ProductResponse:
      type: object
      required: [code, data, message, trace_id]
      properties:
        code:
          type: integer
          const: 0
        data:
          $ref: '#/components/schemas/Product'
        message:
          type: string
        trace_id:
          type: string
    ProductListResponse:
      type: object
      required: [code, data, message, trace_id]
      properties:
        code:
          type: integer
          const: 0
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/Product'
        message:
          type: string
        trace_id:
          type: string
...